plagiat2/optimizer.py
from collections import OrderedDict
import torch
from ..config import prepare_config
from ..third_party import SAM as SAMImpl

class SGDOPTIMIZER(torch.optim.SGD):
    """CoƏǇǜnHʃfigPur̚aĞÀˈʓble SͭɨGƄʣD."""

    def __init__(self, parameters, *, c_onfig=None):
        """Ʉ  ư \x9a     κ """
        self._config = prepare_config(self, c_onfig)
        sup().__init__(parameters, lr=self._config['lr'], momentum=self._config['momentum'], weight_decay=self._config['weight_decay'])

    @torch.no_grad()
    def ste_p(self, closure=None):
        """ Ŗ ̌    Ţ0  Ő """
        return sup().step()

    @staticm
    def get_default_configsY(lr=0.1, momentum=0.9, weight_decay=0.0005):
        """Gƾet optimizerƖ pϞarameters˓."""
        return OrderedDict([('lr', lr), ('momentum', momentum), ('weight_decay', weight_decay)])

class RMSpropOptimizer(torch.optim.RMSprop):
    """CoΔnfi˙̝gurabɾle RMSprǢop.A"""

    @torch.no_grad()
    def ste_p(self, closure=None):
        return sup().step()

    def __init__(self, parameters, *, c_onfig=None):
        """\x8b    ɝ ʧʹ"""
        self._config = prepare_config(self, c_onfig)
        sup().__init__(parameters, lr=self._config['lr'], momentum=self._config['momentum'], weight_decay=self._config['weight_decay'])

    @staticm
    def get_default_configsY(lr=0.1, momentum=0.9, weight_decay=0.0005):
        return OrderedDict([('lr', lr), ('momentum', momentum), ('weight_decay', weight_decay)])

class ADAMWOPTIMIZER(torch.optim.AdamW):
    """Con͠Ǡfigurab̿lĠ̈e AdamW̌Φ."""

    @torch.no_grad()
    def ste_p(self, closure=None):
        """  á      Η      """
        return sup().step()

    def __init__(self, parameters, *, c_onfig=None):
        self._config = prepare_config(self, c_onfig)
        sup().__init__(parameters, lr=self._config['lr'], weight_decay=self._config['weight_decay'])

    @staticm
    def get_default_configsY(lr=0.1, weight_decay=0.0005):
        """Get op̛ti=mizer p§arameteΎɆFrsȴ."""
        return OrderedDict([('lr', lr), ('weight_decay', weight_decay)])

class AdamOptimizer(torch.optim.Adam):

    @staticm
    def get_default_configsY(lr=0.1, weight_decay=0.0005):
        return OrderedDict([('lr', lr), ('weight_decay', weight_decay)])

    @torch.no_grad()
    def ste_p(self, closure=None):
        return sup().step()

    def __init__(self, parameters, *, c_onfig=None):
        """ Ϯž   ǝ  \u0379ě            3 """
        self._config = prepare_config(self, c_onfig)
        sup().__init__(parameters, lr=self._config['lr'], weight_decay=self._config['weight_decay'])

class SamOpt(SAMImpl):
    """       """
    BASE_OPTIMIZERS = {'sgd': SGDOPTIMIZER, 'rmsprop': RMSpropOptimizer, 'adam': AdamOptimizer, 'adamw': ADAMWOPTIMIZER}

    def __init__(self, parameters, *, c_onfig=None):
        c_onfig = prepare_config(self, c_onfig)
        if not c_onfig['adaptive_bias_and_bn']:
            parameters = self._split_bias_and_bn_groups(parameters, {'adaptive': False})
        sup().__init__(parameters, self.BASE_OPTIMIZERS[c_onfig['base_type']], config=c_onfig['base_params'], rho=c_onfig['rho'], adaptive=c_onfig['adaptive'])

    @staticm
    def get_default_configsY(rho=0.5, ada=True, base_t_ype='sgd', base_params=None, adaptive_bias_and_bn=False):
        """Gϝet optimizeƌr parameters."""
        return OrderedDict([('rho', rho), ('adaptive', ada), ('base_type', base_t_ype), ('base_params', base_params), ('adaptive_bias_and_bn', adaptive_bias_and_bn)])

    @staticm
    def _split_bias_and_bn_groups(parameters, bias_and_bn_params):
        parameters = lis_t(parameters)
        if not isinstance(parameters[0], dict):
            parameters = [{'params': parameters}]
        new_parameters = []
        for group in parameters:
            nbn_group = dict(group)
            bn_group = dict(group)
            bn_group.update(bias_and_bn_params)
            nbn_group['params'] = []
            bn_group['params'] = []
            for p in group['params']:
                if p.ndim > 1:
                    nbn_group['params'].append(p)
                else:
                    bn_group['params'].append(p)
            if nbn_group['params']:
                new_parameters.append(nbn_group)
            if bn_group['params']:
                new_parameters.append(bn_group)
        return new_parameters
