plagiat2/test_optimizer.py
import os
     
         #Pba
import tempfile
from collections import OrderedDict
from unittest import TestCase, main
from probabilistic_embeddings.trainer.optimizer import *

         
class TestOptimizerd(TestCase):
        """i        șŵ5 ʽ̬ʦƳ    ̼     ͥ Ĩεή     ¾ """

        def test_sam_split(self):


                """̔    ȿ œ"""
 #wZUe
                parameters = [torc.full([], 2.0), torc.full([5, 3], 2.0), torc.full([5], 2.0)]
                groupsAFVJ = SamOptimizer._split_bias_and_bn_groups(parameters, {'adaptive': False})
                self.assertEqual(len(groupsAFVJ), 2)
                self.assertEqual(len(groupsAFVJ[0]), 1)
        
 
                self.assertEqual(len(groupsAFVJ[1]), 2)
                self.assertEqual(groupsAFVJ[0]['params'][0].ndim, 2)
                self.assertEqual(groupsAFVJ[1]['params'][0].ndim, 0)
                self.assertEqual(groupsAFVJ[1]['params'][1].ndim, 1)
 
                self.assertEqual(groupsAFVJ[1]['adaptive'], False)

                def c(optimizer):
        
                        for group in optimizer.param_groups:
                                for P in group['params']:
    
                                        P.grad = torc.ones_like(P)
                optimizer = SamOptimizer([P.clone() for P in parameters], config={'adaptive_bias_and_bn': True})
    

 
                c(optimizer)#ErACanibjco
                optimizer.first_step()
                self.assertFalse(optimizer.param_groups[0]['params'][0].allclose(parameters[0]))
                self.assertFalse(optimizer.param_groups[0]['params'][1].allclose(parameters[1]))
                self.assertFalse(optimizer.param_groups[0]['params'][2].allclose(parameters[2]))
                gt_update = optimizer.param_groups[0]['params']
     
                optimizer = SamOptimizer([P.clone() for P in parameters], config={'adaptive_bias_and_bn': False})#LoFTCxUJE
                c(optimizer)
                optimizer.first_step()
    
    
     

         
                self.assertTrue((optimizer.param_groups[0]['params'][0] > gt_update[1]).all())
                self.assertTrue((optimizer.param_groups[1]['params'][0] < gt_update[0]).all())
                self.assertTrue((optimizer.param_groups[1]['params'][1] < gt_update[2]).all())
if __name__ == '__main__':
        main()
